+++ CHANGELOG.md
@@ -23,6 +23,10 @@ Classify the change according to the following categories:
     ### Deprecated
     ### Removed
 
+## Develop 
+### Changed 
+- Changed the way we determine which dataset to utilize in the PVWatts API call. Previously, we utilized defined lat-long bounds. Now, we call the Solar Dataset Query API (v2) to determine the dataset to use. 
+
 ## v0.44.0
 ### Added 
 - in `src/settings.jl`, added new const **INDICATOR_COMPATIBLE_SOLVERS**